:recycle: - refactor: move the clear logic into useDataGridProps
